Moved the Navbar to a seperate component
