+++ src/components/Hero.css
@@ -2,7 +2,7 @@
     background-image: linear-gradient(rgba(0,0,0,0.5), rgba(0,0,0,0.5)), url("https://image.tmdb.org/t/p/w1280/1X7vow16X7CnCoexXh4H4F2yDJv.jpg");
     
     height: 85vh;
-    width: 100vw;
+    width: 100dvw;
 
     background-position: center;
     background-repeat: no-repeat;
